backend/requirements.txt
absl-py==2.4.0
Pillow>=10.0.0
asgiref==3.11.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.3
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
protobuf==7.34.0
Pygments==2.19.2
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
tensorflow==2.21.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.2
